requirements.txt
Languages
python

Python Libraries
python-dotenv
google-generativeai
Flask